Add main words ret conds word exclusions
Now it's possible to exclude words from the main words return
conditions: {ANY_MAIN_WORD, "-fast"} --> any main word except "fast"
(this is an example for the reboot command, which was the one that made
me update this).

I also fixed the algorithm that would decide if the conditions agreed
or not with the results. It was broken and I didn't notice until I
reviewed it to add functionality.

And the code that does this is now on WordsVerificationFunction.go.