Increase max sub-commands and add export functions
Now the maximum number of sub-commands is 100 000. Why such a big
number? I've put contacts commands being automatically generated, and
the number is very big (993 on my phone - it's the sum of all the phone
accounts, not just the SIM card or anything). Won't ever go over 10k
surely, but this way, no more issues will appear because of this (unless
10k is not enough for some wow reason). There's also a new exported
constant with the maximum number of sub-commands.

Now there are 2 new functions to manage the commands list. One to add or
update a command, and another one to remove a command. The naming of the
PrepareCmdsArray() function was changed to ReloadCmdsArray().